Move MovetItServer to trollmoves and EventHandler to trollmoves.move_it_base
