+++ trollmoves/move_it_base.py
@@ -30,8 +30,6 @@ import os
 import pyinotify
 from posttroll.publisher import Publisher
 
-from trollmoves.server import EventHandler
-
 LOGGER = logging.getLogger("move_it_base")
 LOG_FORMAT = "[%(asctime)s %(levelname)-8s %(name)s] %(message)s"
 
@@ -141,3 +139,68 @@ def create_publisher(port, publisher_name):
     """Create a publisher using port *port*."""
     LOGGER.info("Starting publisher on port %s.", str(port))
     return Publisher("tcp://*:" + str(port), publisher_name)
+
+
+# Generic event handler
+# fixme: on deletion, the file should be removed from the filecache
+class EventHandler(pyinotify.ProcessEvent):
+    """Handle events with a generic *fun* function."""
+
+    def __init__(self, fun, *args, **kwargs):
+        """Initialize event handler."""
+        pyinotify.ProcessEvent.__init__(self, *args, **kwargs)
+        self._cmd_filename = kwargs.get('cmd_filename')
+        if self._cmd_filename:
+            self._cmd_filename = os.path.abspath(self._cmd_filename)
+        self._fun = fun
+        self._watched_dirs = dict()
+        self._watchManager = kwargs.get('watchManager', None)
+        self._tmask = kwargs.get('tmask', None)
+
+    def process_IN_CLOSE_WRITE(self, event):
+        """On closing after writing."""
+        if self._cmd_filename and os.path.abspath(
+                event.pathname) != self._cmd_filename:
+            return
+        self._fun(event.pathname)
+
+    def process_IN_CREATE(self, event):
+        """On closing after linking."""
+        if (event.mask & pyinotify.IN_ISDIR):
+            self._watched_dirs.update(self._watchManager.add_watch(event.pathname, self._tmask))
+
+        if self._cmd_filename and os.path.abspath(
+                event.pathname) != self._cmd_filename:
+            return
+        try:
+            if os.stat(event.pathname).st_nlink > 1:
+                self._fun(event.pathname)
+        except OSError:
+            return
+
+    def process_IN_MOVED_TO(self, event):
+        """On closing after moving."""
+        if self._cmd_filename and os.path.abspath(
+                event.pathname) != self._cmd_filename:
+            return
+        self._fun(event.pathname)
+
+    def process_IN_DELETE(self, event):
+        """On delete."""
+        if (event.mask & pyinotify.IN_ISDIR):
+            try:
+                try:
+                    self._watchManager.rm_watch(self._watched_dirs[event.pathname], quiet=False)
+                except pyinotify.WatchManagerError:
+                    # As the directory is deleted prior removing the
+                    # watch will cause a error message from
+                    # pyinotify. This is ok, so just pass the
+                    # exception.
+                    pass
+                finally:
+                    del self._watched_dirs[event.pathname]
+            except KeyError:
+                LOGGER.warning(
+                    "Dir %s not watched by inotify. Can not delete watch.",
+                    event.pathname)
+        return
